feat(types): Export types (#8)
This is helpful for annotating helper functions etc. in consumer code.
We're already using these types in our public interface but they're not
named.

Apparently you can't use `export type *` so this just names our existing
types. This might be safer if we add some internal types but it'll make
it easy to forget new types in the future.